requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
Faker==8.0.1
mysql-connector-python==8.0.26
Werkzeug==2.0.1
Flask-Login==0.5.0
python-dotenv==0.19.1
Jinja2==3.0.1
requests==2.26.0